requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
SQLAlchemy==1.4.41
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-json-logger==2.0.4
pytest==7.4.0
playwright==1.44.0
argon2-cffi==23.1.0
httpx==0.24.1
python-multipart==0.0.6
